Added ability to change server parameters
Now the LoginActivity has fields for hostIP and hostPort. If those
fields are left blank, they will default to "localhost" and "8080",
respectively.

I also added a SharedPreferences object that will save your login info
from session to session. This won't be very useful until we implement the
database, but I figured it wouldn't hurt to put it in early.