A multiplicity of changes
Fixed Serializer so that it now correctly deserializes all objects
within the Command objects (some of that code is in Command.java too).
Fixed the ServerPoller so that it actually runs in the background
correctly (CANNOT BE A SINGLETON, do not change it back). Added some
ClientModel.setChanged() calls, which has to precede notifyObservers or
it will not work correctly. Fixed the layout for the GameListActivity.
And probably made other changes, but I don't want to elucidate any more.